% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aninet.R
\docType{package}
\name{aninet}
\alias{aninet}
\title{aninet: Statistical Models for Animal Social Networks}
\description{
The aninet package provides functionality to fit several kinds of statistical model to animal social data.
The primary functionality is the calculation of social differentiation, and generalised affiliation indices.
Also included is an implementation of the generalised linear model quadratic assignment procedure, useful for regression
applications in social networks.
}
