% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_differentiation.R
\name{association_complexity}
\alias{association_complexity}
\title{Complexity of social associations using mixture models}
\usage{
association_complexity(
  Den,
  Num,
  maxJ = 9,
  criterion = "ICL",
  maxrep = 20,
  tolfun = 1e-06,
  minprior = 0,
  maxiter = 1000
)
}
\arguments{
\item{Den}{Numeric, denominator of association indices}

\item{Num}{Numeric, numerator of association indices}

\item{maxJ}{Numeric, maximum number of components to fit}

\item{criterion}{Character, information criteria to use for model selection.}

\item{maxrep}{Numeric, maximum number of repitions for fitting process}

\item{tolfun}{Numeric, tolerance for stopping criteria}

\item{minprior}{Numeric, minimum portion of dyads in a cluster}

\item{maxiter}{Numeric, maximum number of iterations for convergence}
}
\value{
A named list with two elements. The \code{summary} element contains a summary of all fit models, including number of components, estimated complexity, and information criteria. The \code{best_model} contains more detailed information on the model chosen by the specified information criterion. This element is a named list, with detailed information on the number of components, component means and frequencies, estimated complexity, likelihood, and information criteria.
}
\description{
Fits a binomial mixture model to observed associations and calculates the resulting complexity measure.
}
\details{
Social complexity can be defined in a number of ways. This algorithm seeks to measure the diversity of "classes" of assocation occuring at any point. The model describes the association data as a mixture of K association types, each with their own probability of association.
By fitting models with K = {1, 2, ..., maxJ}, we can perform model selection to choose the "best" number of relationship types. Each type has a fitted association probability (p) and portion of dyads (a). The portion of association of each type is then q = p*a, and we then measure the entropy of q to arrive at a measure of complexity.
The model additionally fits beta-binomial models to derive the dispersion within components.
Model selection can use any of AIC, BIC, or ICL to decide on the best model. In simulations ICL performs most consistently, but may underestimate complexity when sample sizes are small.
}
