% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{association_hclust}
\alias{association_hclust}
\title{Hierarchical clustering of association indices using modularity}
\usage{
association_hclust(network, method = "average")
}
\arguments{
\item{network}{A symmetric square matrix of association indices}

\item{method}{The linkage method for hierarchical clustering}
}
\value{
A named list containing the best modularity ("maximum_modularity"), the cophenetic clustering coefficient ("CCC"), the clustering tree ("tree"), a dataframe of cut heights and modularity ("cuts"), and merge information ("merge").
See \code{hclust} for details on manipulating and plotting trees.
}
\description{
An implementation of the network clustering method described by Lusseau (2007) and implemented in SOCPROG.
}
\details{
This method works by transforming the association indices into distances (by subtracting them from 1), running hierarchical clustering as usual, and then calculating the modularity of the division at each split.
This method is currently only designed to be used with association-like indices (which have a maximum of 1).
}
