% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gais.R
\name{attribute_similarity}
\alias{attribute_similarity}
\title{Attribute similarity matrices}
\usage{
attribute_similarity(attr, type = "discrete")
}
\arguments{
\item{attr}{Vector, individual attributes}

\item{type}{Character, type of similarity to calculate}
}
\value{
A square matrix of similarity values
}
\description{
Generate square matrices of similarity between individuals from individual attribute data
}
\details{
The available types of similarity are \code{"discrete"} for 1-0 similarity, \code{"absdiff"} for negative absolute difference, and \code{"sqdiff"} for negative squared difference. If the attribute data is not numeric, the function will default to \code{"discrete"}.
}
