% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_csi.R
\name{dyadic_csi}
\alias{dyadic_csi}
\title{Calculate composite sociality indices from a set of social networks}
\usage{
dyadic_csi(networks)
}
\arguments{
\item{networks}{A list of networks from which to calculate composite indices. Must be of the same dimensions, and in the same order.}
}
\value{
A square matrix of dyadic compisite sociality indices
}
\description{
Collapses multiple types of interaction or association into a single network of composite indices of social relationships
}
\details{
Dyadic composite sociality indices are calculated as in Sapolsky et al. (1991).
The dyadic CSI is calculated as:

\deqn{CSI_{ij} = frac{sum_{d = 1}^{D} frac{f_{dij}}{\bar{f_d}} } {D}}

Before calculating these indices, it is important to consider whether the included networks positively covary. For example, grooming and proximity may be useful for generating a composite index, but it likely does not make sense to combine grooming and aggression.
}
