% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbi_null_models.R
\name{extend.gbi_MCMC}
\alias{extend.gbi_MCMC}
\title{Extend a MCMC Run for Group-By-Individual Permutations}
\usage{
extend.gbi_MCMC(x, samples, ...)
}
\arguments{
\item{x}{A \code{gbi_null} object as generated by \code{gbi_MCMC}}

\item{samples}{The number of additional samples to draw}

\item{...}{Additional arguments to generate the test statistic}
}
\description{
Runs additional permutation steps from a \code{gbi_MCMC} run.
}
\details{
This function will extend the sampling of a \code{gbi_null} object. All settings, including the thin and statistics being calculated, will be the same as the initial run.
}
