% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gais.R
\name{gai}
\alias{gai}
\title{Generalised Affiliation Indices}
\usage{
gai(
  formula,
  family = "gaussian",
  weights = NULL,
  offset = NULL,
  type = c("deviance", "pearson", "response")
)
}
\arguments{
\item{formula}{A glm style formula giving the response and predictor matrices.}

\item{family}{Error family for fitting, either a \code{glm} family or one of \code{"betar"} or \code{"negbin"}.}

\item{weights}{Weights used for fitting, see Details.}

\item{offset}{Offset used in fitting, see Details.}

\item{type}{Type of residuals to calculate}
}
\value{
A square matrix containing the generalised affiliation indices.
}
\description{
Caluclates a social network based on generalised affiliation indices (GAI). When non-social factors influence the observed interactions or associations, using GAIs can help to remove the effect of these factors, revealing underlying social affinities. This is done by fitting a generalised linear model, with the observed network as a response and non-social confounds as predictors, and using the residuals as a measure of affiliation.
}
\details{
For association data, \code{family = "binomial"} is the most appropriate. In this case, the observed association indices should be the response matrix, and the dyadic denominators should be passed to the \code{weights} argument.
For interaction rates, one of \code{"poisson"}, \code{"quasipoisson"}, or \code{"negbin"} should be passed to \code{family}, with log(sampling effort) passed as an offset.
For completeness, beta regression is also included, using \code{betareg::betareg()}.
Generally, its a good idea to test whether the predictors are important prior to calculating GAIs. This can be done using GLMQAP.
}
