% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{get_denominator}
\alias{get_denominator}
\title{Get denominator of association indices}
\usage{
get_denominator(data, index = "SRI", data_format = "SP", return = "vector")
}
\arguments{
\item{data}{The association data.}

\item{index}{The association index to use, one of \code{"SRI"}, \code{"HWI"}, and \code{"BII"}.}

\item{data_format}{Character, indicating what form the data is in.}

\item{return}{Character, indicating whether a vector or square matrix should be returned}
}
\value{
Either a vector of denominators (if \code{return = "vector"}) or a square matrix of denominators (if \code{return = "matrix"}).
}
\description{
For many functions, the denominator of association indices is needed, rather than the final network. This function allows to user to easily extract these.
}
\details{
Data format can be either \code{"SP"} or \code{"GBI"}, the two main data formats used by \code{"asnipe"}. Passing \code{index = "SRI"} returns the denominator of the simple ratio index, \code{index = "HWI"} returns half-weight index, while \code{index = "BII"} uses a "both-identified" index.
}
