% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{get_numerator}
\alias{get_numerator}
\title{Get numerator of association indices}
\usage{
get_numerator(data, data_format = "SP", return = "vector")
}
\arguments{
\item{data}{The association data}

\item{data_format}{Character, indicating what form the data is in.}

\item{return}{Character, indicating whether a vector or square matrix should be returned}
}
\value{
Either a vector of numerators (if \code{return = "vector"}) or a square matrix of numerators (if \code{return = "matrix"}).
}
\description{
For many functions, the numerator of association indices is needed, rather than the final network. This function allows to user to easily extract these.
}
\details{
The association index does not need to be specified, as they almost all use the same numerator. Data format can be either \code{"SP"} or \code{"GBI"}, the two main data formats used by \code{"asnipe"}.
}
\examples{
get_numerator(srkw_sightings,data_format = "GBI", retunr = "vector")
}
