% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gais.R
\name{joint_gregariousness}
\alias{joint_gregariousness}
\title{Joint gregariousness}
\usage{
joint_gregariousness(network, log.jg = T)
}
\arguments{
\item{network}{A square matrix, defining the network from which to calculate joint gregariousness}

\item{log.jg}{Logical, indicating whether the log-transformed scores are to be returned.}
}
\value{
A square matrix of dyadic joint gregariousness scores.
}
\description{
This function calculates the "joint gregariousness" for each dyad, as given by Whitehead & James (2015)
}
\details{
Joint gregariousness is defined by Whitehead & James (2015) as:
\deqn{joint gregariousness[i,j] = log(greg[i] * greg[j])}, where \eqn{greg[i]} is \deqn{\sum_{k \neq j} network[i,k]}
This definition removes each dyads edge weight from the calculation of their joint gregariousness.
The function allows the user to get the non log-transformed value. This is useful if there are individuals with only as single non-zero edge weight.
}
