% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbi_null_models.R
\name{plot.gbi_null}
\alias{plot.gbi_null}
\title{Plot Null Distributions for Group-By-Individual Matrices}
\usage{
\method{plot}{gbi_null}(x)
}
\arguments{
\item{x}{A \code{gbi_null} object as produced by \code{gbi_MCMC}}
}
\description{
Diagnostic plots for GBI null hypotheses tested with the Manly-Bejder MCMC procedure
}
\details{
This function will first plot the trace plots for the null distributions of each test statistic. It will then plot the change in estimated p-values over the chains.
}
