% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_space.R
\name{plot.latsoc}
\alias{plot.latsoc}
\title{Plotting latent space fits}
\usage{
\method{plot}{latsoc}(
  object,
  post.method = "mean",
  edge.col = NULL,
  edge.lwd = NULL,
  vertex.pch = 19,
  vertex.col = "black",
  vertex.cex = 1,
  labels = F,
  label.col = NULL,
  label.cex = NULL,
  ...
)
}
\arguments{
\item{object}{A model returned by \code{latent_space}}

\item{post.method}{Character specifying the method for estimating the posterior latent positions. One of "mean", "median", or "mds", the latter of which uses multidimensional scaling.}

\item{edge.col}{Square matrix indicating the colors to plot edges. If NULL, edges are plotted in black, with transparency determined by their edge weight.}

\item{edge.lwd}{Square matrix indicating edge line widths. If NULL, edge widths are proportional to edge weight.}

\item{vertex.pch}{Plotting characters for vertices/nodes.}

\item{vertex.col}{A vector indicating the colors for plotting nodes.}

\item{vertex.cex}{Expansion factor for vertices/nodes.}

\item{labels}{Logical, should node labels be plotted?}

\item{label.col}{Colors for vertex labels.}

\item{label.cex}{Numeric, character expansion factor for labels.}

\item{...}{Additional arguments to be passed to \code{plot}}
}
\description{
Plots the fitted network object with node positions determined by latent space modelling. 1 and 2 dimensional fits can be plotted directly, while > 3 dimensional fits use multidimensional scaling to project distances into 2 dimensions.
}
\details{
This function uses base R plotting methods. More advanced network plotting methods are available (e.g. through the ggraph package), but this interface provides an easy way to visualize the results of the model quickly.
}
