% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmqap.R
\name{print.glmqap}
\alias{print.glmqap}
\title{Print function for glmqap objects}
\usage{
\method{print}{glmqap}(x)
}
\arguments{
\item{x}{A glmqap object}
}
\value{
Printed results
}
\description{
Prints summary information about GLMQAP models
}
\details{
Prints formatted results from GLMQAP fits, including formula, family, permutations, weights and offsets, coefficients, standard errors, test statistics, and p-values
}
