% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_attributes}
\alias{srkw_attributes}
\title{Individual attributes of southern resident killer whales}
\format{
A data.frame containing individual ID, age, sex, age-sex class, and matriline identity.
}
\usage{
srkw_attributes
}
\description{
Contains individual information about the members of J pod in 2019
}
\keyword{datasets}
