% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_codetection}
\alias{srkw_codetection}
\title{Dyadic codetections of southern resident killer whales}
\format{
A symmetric square matrix containing the co-detection measures.
}
\usage{
srkw_codetection
}
\description{
Portion of sampling time in which each pair of whales in J pod was observed together. A value of 1 would indicate whales were always seen together, while 0 would indicate they were never seen together.
}
\keyword{datasets}
