% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_contact}
\alias{srkw_contact}
\title{Contacts between southern resident killer whales}
\format{
A symmetric square matrix, with rows and columns representing individuals, and the entried indicating the contact rates between them.
}
\usage{
srkw_contact
}
\description{
A dataset containing the number of contacts observed between members of J pod
Data were collected from an unmanned aerial vehicle over 11 hours of observation in 2019
}
\keyword{datasets}
