% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_kinship}
\alias{srkw_kinship}
\title{Estimated maternal kinship between southern resident killer whales}
\format{
A symmetric square matrix containing pairwise relatedness estimates
}
\usage{
srkw_kinship
}
\description{
Pairwise relatedness coefficients between individual southern resident killer whales in J pod, based on maternal pedigrees.
}
\keyword{datasets}
