% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_sampling}
\alias{srkw_sampling}
\title{Dyadic sampling effort of southern resident killer whales}
\format{
A symmetric square matrix containing the dyadic sampling efforts
}
\usage{
srkw_sampling
}
\description{
Total observation time (in seconds) for each pair of whales in J pod during which interactions were observed.
}
\references{
Weiss, M.N., Franks, D. W., Giles, D. A., Youngstrom, S., Wasser, S. K., Balcomb, K. C., Ellifrit, D. K., Domenici, P., Cant, M. A., Ellis, S., Nielsen, M. L. K., Grimes, C., & Croft, D. P. (In press) Age and sex influence social interactiosn, but not associations, within a killer whale pod. Proceedings of the Royal Society B, In press.
}
\keyword{datasets}
