% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_sightings}
\alias{srkw_sightings}
\title{Sightings of southern resident killer whales over 10 days}
\format{
A day-by-individual matrix, with 1 indicating an individual was seen on a given day, and a 0 indicating they were not.
}
\usage{
srkw_sightings
}
\description{
Sightings history of J pod over a 10 day observation period
}
\references{
Weiss, M.N., Franks, D. W., Giles, D. A., Youngstrom, S., Wasser, S. K., Balcomb, K. C., Ellifrit, D. K., Domenici, P., Cant, M. A., Ellis, S., Nielsen, M. L. K., Grimes, C., & Croft, D. P. (In press) Age and sex influence social interactiosn, but not associations, within a killer whale pod. Proceedings of the Royal Society B, In press.
}
\keyword{datasets}
