% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_surfacing}
\alias{srkw_surfacing}
\title{Synchronous surfacings between southern resident killer whales}
\format{
A symmetric square matrix containing the observed number of synchronous surfacing events between each dyad.
}
\usage{
srkw_surfacing
}
\description{
Number of synchronous surfacing events between members of J pod in the summer of 2019.
}
\keyword{datasets}
