% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbi_null_models.R
\name{summary.gbi_null}
\alias{summary.gbi_null}
\title{Summarise results of Group-By-Individual Randomisations}
\usage{
\method{summary}{gbi_null}(x)
}
\arguments{
\item{x}{a \code{gbi_null} object as produced by \code{gbi_MCMC}}
}
\description{
Generate relevant summaries of null hypothesis testing using MCMC randomisations of group-by-individual matrices
}
\details{
The function first produces a summary of the null distribution, with mean, median, SD, confidence intervals, and diagnostics (effective sample size and potential scale reduction factor).
The function then shows estimated upper, lower, and two-tailed p-values of the observed values against the null. The output includes confidence intervals for the p-values, which are estimated from the effective sample size using binomial theory.
}
