% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_differentiation.R
\name{social_differentiation}
\alias{social_differentiation}
\title{Estimation of social differentiation}
\usage{
social_differentiation(Num, Den, initial.params = c(0.1, 0.1), nsim = 1e+05)
}
\arguments{
\item{Num}{Numeric vector, numerator of the association indices}

\item{Den}{Numeric vector, denominator of association indices}

\item{initial.params}{Initial parameters for model fitting}

\item{nsim}{Number of parametric bootstraps to draw for error estimation. See details.}
}
\value{
A matrix containing the estimated social differentiation, the CV of the observed associations, and the estimated correlation between true and observed association indices, along with standard errors and confidence intervals.
}
\description{
Estimates the coefficient of variation of the underlying association probabilities using maximum likelihood, assuming underlying association probabilities follow a beta distribution.
}
\details{
Social differentiation is commonly defined as the coefficient of variation of the true, underlying association probabilities. This estimation procedure assumes that the underlying probabilities follow a beta distribution, and estimates the parameters of this distribution given the observed association indices.
The estimation of social differeniation serves as both a helpful descriptor of social structure, and a useful measure for determining the power and precision of social analyses. The estimated correlation between true and observed association indices can be derived by dividing the estimated social differentiation by the observed CV of association indices.
In some edge cases, where social differentiation is very high, a correlation greater than 1 can be estimated. In these cases we recommend primarily interpretting the lower bound of the confidence interval.
Results using this package may differ slightly from results obtained using SOCPROG in MATLAB; this is because SOCPROG uses a slightly different method to estimate the likelihood, numerically integrating over the possible probability values given the parameters of the beta distribution. In this package, we explicitly use the probability mass function of the beta-binomial distribution.
Differences should be small, however results from this function should be more precise than those obtained in SOCPROG.
This function estimates the standard error and confidence intervals for both social differentiation and the estimated correlation. This is based on a parametric bootstrap from the estimated mean and covariance matrix for the underlying beta distribution parameters.
A potentially more conservative way to estimate this error is through a non-parametric bootstrap or jackknife of the raw association data. We recommend setting nsim = 1 if calculating social differentiation on many resampled datasets, to speed up computation.
}
\examples{
X <- get_numerator(srkw_sightings, return = "vector", data_format = "GBI")
D <- get_denominator(srkw_sightings, return = "vector", data_format = "GBI")
social_differentiation(X, D, method = "Beta-binomial")

}
