% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_differentiation.R
\name{social_differentiation}
\alias{social_differentiation}
\title{Estimation of social differentiation}
\usage{
social_differentiation(
  Num,
  Den,
  method = c("Whitehead", "Beta-binomial"),
  res = 0.001,
  initial.params = c(0.1, 0.1)
)
}
\arguments{
\item{Num}{Numeric vector, numerator of the association indices}

\item{Den}{Numeric vector, denominator of association indices}

\item{method}{Character, indicating what likelihood function to use}

\item{res}{Resolution used for integration (only used for \code{method = "Whitehead"})}

\item{initial.params}{Initial parameters for model fitting}
}
\value{
A matrix containing the estimated social differentiation, the CV of the observed associations, and the estimated correlation between true and observed association indices, along with standard errors and confidence intervals.
}
\description{
Estimates the coefficient of variation of the underlying association probabilities using maximum likelihood, assuming underlying association probabilities follow a beta distribution.
}
\details{
Social differentiation is commonly defined as the coefficient of variation of the true, underlying association probabilities. This estimation procedure assumes that the underlying probabilities follow a beta distribution, and estimates the parameters of this distribution given the observed association indices.
The estimation of social differeniation serves as both a helpful descriptor of social structure, and a useful measure for determining the power and precision of social analyses. The estimated correlation between true and observed association indices can be derived by dividing the estimated social differentiation by the observed CV of association indices.
This function can perform estimation in two ways. The first is using the likelihood function given by Whitehead (2009), using \code{method = "Whitehead"}. The second uses the probability mass function of the beta-binomial distribution using \code{method = "Beta-binomial"}.
Because it requires numeric integration, the \code{"Whitehead"} method is much slower than the \code{"Beta-binomial"} method, however has the upside of being directly comparable to results from SOCPROG (and thus the vast majority of values reported in the literature).
Data should be numeric vectors of numerators and denominators, rather than square matrices. See \code{get_numerator()} and \code{get_denominator()} for easy ways to extract these from raw data.
This function calculates and returns the standard error and confidence interval for the estimated social differentiation and correlation. This is based on the inverse Hessian matrix from the optimization routine. Note that a bootstrap of the raw data may be more robust.
}
\examples{
X <- get_numerator(srkw_sightings, return = "vector", data_format = "GBI")
D <- get_denominator(srkw_sightings, return = "vector", data_format = "GBI")
social_differentiation(X, D, method = "Beta-binomial")

}
