% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_attributes}
\alias{srkw_attributes}
\title{Individual attributes of southern resident killer whales}
\format{
A data.frame containing individual ID, age, sex, age-sex class, and matriline identity.
}
\usage{
srkw_attributes
}
\description{
Contains individual information about the members of J pod in 2019
}
\references{
Weiss, M.N., Franks, D. W., Giles, D. A., Youngstrom, S., Wasser, S. K., Balcomb, K. C., Ellifrit, D. K., Domenici, P., Cant, M. A., Ellis, S., Nielsen, M. L. K., Grimes, C., & Croft, D. P. (In press) Age and sex influence social interactiosn, but not associations, within a killer whale pod. Proceedings of the Royal Society B, In press.
}
\keyword{datasets}
