% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_kinship}
\alias{srkw_kinship}
\title{Estimated maternal kinship between southern resident killer whales}
\format{
A symmetric square matrix containing pairwise relatedness estimates
}
\usage{
srkw_kinship
}
\description{
Pairwise relatedness coefficients between individual southern resident killer whales in J pod, based on maternal pedigrees.
}
\references{
Weiss, M.N., Franks, D. W., Giles, D. A., Youngstrom, S., Wasser, S. K., Balcomb, K. C., Ellifrit, D. K., Domenici, P., Cant, M. A., Ellis, S., Nielsen, M. L. K., Grimes, C., & Croft, D. P. (In press) Age and sex influence social interactiosn, but not associations, within a killer whale pod. Proceedings of the Royal Society B, In press.
}
\keyword{datasets}
