% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_sampling}
\alias{srkw_sampling}
\title{Dyadic sampling effort of southern resident killer whales}
\format{
A symmetric square matrix containing the dyadic sampling efforts
}
\usage{
srkw_sampling
}
\description{
Total observation time (in seconds) for each pair of whales in J pod during which interactions were observed.
}
\keyword{datasets}
