% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_sightings}
\alias{srkw_sightings}
\title{Sightings of southern resident killer whales over 10 days}
\format{
A day-by-individual matrix, with 1 indicating an individual was seen on a given day, and a 0 indicating they were not.
}
\usage{
srkw_sightings
}
\description{
Sightings history of J pod over a 10 day observation period
}
\keyword{datasets}
