% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srkw_surfacing}
\alias{srkw_surfacing}
\title{Synchronous surfacings between southern resident killer whales}
\format{
A symmetric square matrix containing the observed number of synchronous surfacing events between each dyad.
}
\usage{
srkw_surfacing
}
\description{
Number of synchronous surfacing events between members of J pod in the summer of 2019.
}
\references{
Weiss, M.N., Franks, D. W., Giles, D. A., Youngstrom, S., Wasser, S. K., Balcomb, K. C., Ellifrit, D. K., Domenici, P., Cant, M. A., Ellis, S., Nielsen, M. L. K., Grimes, C., & Croft, D. P. (In press) Age and sex influence social interactiosn, but not associations, within a killer whale pod. Proceedings of the Royal Society B, In press.
}
\keyword{datasets}
